project10/pyproject.toml
[tool.poetry]
name = "hw-project10"
version = "0.1.0"
description = ""
authors = ["GievskiyIgor <121798393+GievskiyIgor@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
redis = "^5.0.6"
django-environ = "^0.11.2"
environ = "^1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
